seperate out cart related models from merchi.js
this completes the moving out of model definitions from that file.
